requirements.txt
fastapi[standard]
cryptography
PyJWT
httpx
